refactor: name PDFs from their URL; move URL helpers to _content_file
content_pdf_url() named every file from a process-global counter, so
`file_00N.pdf` depended on how many earlier calls happened in the process.
That makes request bodies non-deterministic (it broke a body-matched VCR
cassette purely on test ordering) and throws away a real name that's already
in the URL. Now the last path segment wins when it ends in .pdf, with the
counter as a fallback -- matching what content_document_url() already did.

Reaching filename_from_url() from _content_pdf.py would have been circular
(_content_document already imports parse_data_url from there), so both URL
helpers move to _content_file.py, the module this branch added for logic
shared by ContentPDF and ContentDocument. That also stops _content_image.py
and _inspect.py from importing a URL helper out of the PDF module.
